+++ tests/BadRequestHandlerTest.php
@@ -11,23 +11,27 @@ use Psr\Http\Message\ServerRequestInterface;
 use Yiisoft\Http\Method;
 use Yiisoft\Http\Status;
 use Yiisoft\Request\Body\BadRequestHandler;
+use Yiisoft\Request\Body\ParserException;
 
 final class BadRequestHandlerTest extends TestCase
 {
-    public function testShouldReturnCode400(): void
+    public function testShouldReturnCode400AndCorrectErrorInBody(): void
     {
         $response = $this
             ->createHandler()
             ->handle($this->createRequest());
         $this->assertEquals(Status::BAD_REQUEST, $response->getStatusCode());
+        $this->assertEquals(Status::TEXTS[Status::BAD_REQUEST], (string)$response->getBody());
     }
 
-    public function testShouldReturnCorrectErrorInBody(): void
+    public function testShouldReturnCode400AndExceptionMessageInBody(): void
     {
+        $e = new ParserException('JGURDA');
         $response = $this
             ->createHandler()
-            ->handle($this->createRequest());
-        $this->assertEquals(Status::TEXTS[Status::BAD_REQUEST], (string)$response->getBody());
+            ->handle($this->createRequest(), $e);
+        $this->assertEquals(Status::BAD_REQUEST, $response->getStatusCode());
+        $this->assertEquals(Status::TEXTS[Status::BAD_REQUEST]. "\n" . $e->getMessage(), (string)$response->getBody());
     }
 
     private function createHandler(): BadRequestHandler